restructure search folder
